make jms serializer leave properties alone
This results in the properties being mapped with an uppercase first char (ie showInMenu will be ShowInMenu). This is actually closer to what we want since it used to create 'show_in_menu'. It seems that we would need to patch jms/serializer if we would want to make it serialize stuff as 'showInMenu'.