bugfix: bundle-bundle config/routing loader
The loader was not loadig any files and this lead to composer breaking
badly in rather normal circumstances. Graviton bundles now need to
specify exactly what they need loaded. The bundle might look like
follows to be considered.

````javascript
{
  "extra": {
    "graviton-config-import": "path/to/config.xml",
    "graviton-routing-import": "path/to/routes.xml"
  }
````

I'm putting this directly into master since it seems rather critical.

I'm also adding a version to composer.json and bumping that to 0.1.1.
This is a semver.org version and the logic of the release history up
until now is like so:

* 0.0.0 = what we had in svn
* 0.0.1 = inital github version
* 0.1.0 = bundle-bundle feature
* 0.1.1 = this fix