add guzzle bundle support
I had originally planned to use this for the acceptance test client and I'm adding
it even tough I've given up that Idea. Turns out we need closer integration than
guzzle can offer for that. I'm adding guzzle now since I expect that we will be
using it anyway as soon as we need to also do RESTful requests.