generate routes automagically for /core
This needed quite some trickery but it seems to do what I want.

For some reason I had it in my head that decicion whether to
generate a route or not must be triggered by tagging a controller
in the service container. As it turned out, this was harded to
accomplish than I had orignally anticipated.