add canonical url for schemata
These should be the key in a clients schema cache. This was implemented due
to the fact that there is alot of redundancy in the OPTIONS responses from all
the items. I still need to figure a clean way that client that already have the
canonical results cached may find a way to not get a body in OPTION responses.