add translatable property on schema properties
This is a hack to make it easier for the backend to recognize
i18nized fields that need special handling. I'm extending json
schema like so since i could not find anything resembling a
decent standard that supports this. I'll need to take things like
like this into account when I start validating my schemas against
the yet to be written v5 draft.