rename TaxonomyBundle to EntityBundle
I'm doing this in preparation for additional services and due to the fact that countries are a
type entity and not a taxon anyhow. I might create a new TaxonomyBundle if something
more taxon like show up at a later stage. I'm thinking this com kind of classification tree,
maybe even stuff like ``/core/product``.