create document using simple template
This is in lieue of something more fancy akin to the doctrine-odm DocumentGenerator Tool. It still needs some work to
fully support the same set of features but I think the code will be more manageable this way. I'll probably need to make
sure that devs get useable extension points in here somewhere so we can always regenerate a given document but
my just keep using doctrine:mongodb:generate:documents as before when doing edits.