implement etag and if-none-match listeners
For now using the sha1 sum of the response should suffice for our etag generation.

If we ever run in to performance problems due to this i might consider storing a randomly generated
ETag alongside each document that needs caching. Chanches are this will get revisited as soon as
we have more data in our systems and we run into performance issues.