throw at least some errors
These don't change the fact that we need to implement some kind of validation
and logging in these areas of our code.

If one happens to have a small error in a definition it should not take that much
debugging to figure it out :)

I think we already have a GRV on this, for now this is what happend due to a typo
of mine that i realized to late.