refactor schema aware listeners
This should help a bit with our phpunit performance issues. Some of them seem to stem from the
fact the container builder does multiple passes on ContainerAware things. This gets rid of a small
part of them. It won't make phpunit fast, but it is a step in the right direction.

File this under cleaning up legacy code from the olden days when injecting the container was still
the rage.