basic writing of mongodb-references as $ref for /core/module
This isn't quite the fork of jms_serializer we had expected (yet?). While there are still some kinks to work out, It looks like
a SerializerHandler is the way to go.

This works for pushing out mongodb-references as $ref (see /core/module for an example). There is still some hardcoded stuff
and magic in it that I expect to remove over the next few pushes by refactoring it into sane stuff like container tags and configuration.