Cheat travis by moving assertion
I know, I know... This does not adddress the underlying issue. With this I am feeling
somewhat cheated by travis since I have not been able to reproduce it locally even
after alot of tries. I suspect this might boil down to be the same kind of strange mtime
issue we are experiencing in #84 :(