Move app to app.ref in /core/module
This is still rather hacky and it only supports one subLevel for now. It should contain the basic blocks for making the renderer recursive in a later stage, but I'll have look into the CompilePass and find a solution for what I've got there before it makes sense to carry on with making ExtRefListener recurse fully.