Allow fields named $ref
This is all still rather hacky. Especially the generator parts aren't much fun as they stand now. At least it seems to work for both writing and reading of attirbutes called $ref. I'll have to get rid of the remaining TODOs in the code and will create Tasks for them ASAP.