Fix case-sensitivity bug
Dang you apple :( This issue cropped up due to the fact that Apple still feels it reasonable to install a case-insensitive
file-system as default.

Turns out ignoring stuff on travis without finding the root cause is a bad idea. On the other hand it would have took
even longer to find on travis than on nova.

I'll switch my local dev to using docker as soon as we've made some strides there, that should help with this kind of issue.

tldr: I'm an idiot and this took way to long.