Don't load JsonDefs from Mongo
Thes removes the sillyness that was loading definitions from mongodb during build time. I simply introdued way to much side effects during install.

The Defs that where loaded this was are now directly in a git repo where we have control of when and how they change and we always know which version was used at the time.

The new grv/evoja-loadconfig-bundle has more info on how to use them and how to update the defs in the bundle.

Since this also decouples our build process from injected loadconfigs via mongo it will make automated deploys of loadconfig based wrappers possible.