Move config to parameters.yml.dist properly
This add some indirection to the s3 config and allows us to specify it either manually during install or through cloudfoundry.

We'll need to take care of letting customers that don't use s3 override this in a way that matches their use-case soonish.

For now I'm implementing it like so...