Fix app controller test by splitting assertion
This seems to have changed in php and indicates that travis updated their stack :)

For some completely unrelated reason I also needed to change the 500 assertion to a 400 assertion. I'm not sure as to why this was needed, but a client posting malformed JSON should clearly not be getting sent a 500 header. The 400 that is replacing this is actually a much more concise message for these purposes.