Switch to constructor injection, add ifnal keyword and don't use container in most cases
We still need to get rid of container use in getControllerFromRoute. I might do a factory on that but have not really decided yet. I'm pretty sure this is also done in a similar fashion elsewhere so that adding such a factory will probably have other implications elsewhere as well.