Move mainf finding and loading of config to trait
I'll use this trait in a new CompilerPass that aalso needs to build maps based on config files but does not want to
map them exactly the same way. The trait here actually has an abstract function users need to implement that does
the actual loading and recursing over the doctrine config files. In itself the trait does not touch the map.