Use strategy=set on all many references
We'll need to look and see if this generates the BSON we need. If it doesn't we can still try witching to setArray. I think set should do since we should not have any numerical arrays that are not sequential as far as I can tell.

It also looks like this is still leaving us with the edge case of not having a way to completely amy a collection. If I update a collection i can change individual values and even remove values. As soon as I remove all of them at once (like the test removes the one value) the collection doesn't get touched. I stil need to figure out how doctrine/mongodb-odm expects us to make it remove whole collections.