Redo self and paging links
The only char that is still encoded is , (encoded as %2C). This is due to the fact that most apps don't seem to do more than just explode link headers on the , char.

For all other cases the rql part in the headers in now encoded just like it was in the request. Basically the only normalization that is done on the query string is what has already here and encoding of ,.