Make service.$ref optional
The "Product" service that was being referenced is very old and probably not in a useable state. We might want to switch these to being strings and not using the internal reference mongoref-storage in some cases. I see some use-cases in here that should probably be addressed so this really does what graviphoton wants.