Add rql param to swagger.json
We need to do some small changes in swagger-ui for this to work. I haven't found a swagger way to specify that a param replaced the full query string.

In theory we could generate spec for heaps of FIQL style params, in the end we might as well start forking swagger-ui proper soon anyhow...