Split into unit and integration tests
Well, sort of at least. With this change all the different WebTestCase based tests are run in the "integration" suite and the remainder in the "unit" suite.

Striclty speaking the both do integration testing, the unit suite has heaps of database fixtures and is dependant on mongodb just as the "integration" suite as well.