Replace recordGetter with refField
I'll use the id from refField as key value when doing output. I'm removing the recordGetter feature since I think we should only use id here.

It'll be worth more to clean up /core/module to cope with this than to do anything else. It's not like a configuration will have to use all the available modules if it doesn't want to.