Allow defining of mongo db collection a service maps to
This allows us to define what collection a service takes its data from. This can be used if the same data need to be presented differenty depending on the endpoint.

I plan on using it to keep an editable version of a service containing some dynamic keys available at a different path. While it won't be used in that way as soon as dynamic keys are writable I still think this might have it's merits later on.

It can also be used if we want to reference something from an existing collection in a readonly service so there's that as well.