Only check getenv() return for !== 'false'
Since getenv() will return a boolean false when a var has not been set we need to allow that and fall back to to using composer per default.

The file writing should only be disabled if the env variable has explicitly been set to false which would get return as the string 'false' by getenv().