Hide id in embedded documents
This introduces the realId field which has the sole purpose of making it possible for embedded-objects to have a id field as mandated by doctrine odm. The realId field is then hidden from users by stripping it in doctrine.

This is implemented rather (quick) and dirty due to the generator being rather nasty.