Hide realId from schema output
This is a rather quick solution. I tried removing the realId earlier, but need to do it so late due so I don't break everywhere else this is used.

I'm also checking that the field is an id since that enables us to still use the name realId for other purposes if we need to.