Simple migration for translatable index
I prepared this so it will work in wrappers and you need to do some small changes if you
want to test it without cloning it in a wrappe (as noted in migrations.yml). We should not
really need this on bare graviton deploys anyhow since those are not getting a lot of action
and have way less potential for conflicts.