Remove shunt 3rdparty config
We do not seem to need this configured properly for regular testing. We should probably invest in an environment that helps us test these better though. I'm mainly deactivating it to make our tests run faster as this was broken wehn i deactivated kernel.debug on all tests for some reason.