Add target.indexes from json defs to doctrine mapping
By configuring doctrine properly the indexes will get created if they are defined in the def. Since we create the indexes in the same way ml3k does, this should create exactly the same indexes and not lead to any index conflicts. For that reason it's also left up to the mongodb driver to name the index.