Downgrade symfony/symfony by blocking 2.8.9
We had issues with OPTIONS requests being answered like they where GET requests and could ony narrow them down to the release that is being blocked by this commit. The issues we where seeing only affected SYMFONY_ENV=prod which is why we are rolling back the update thus letting us check for the real reason without a broken master.